move checks for ref-maybe-const on forall earlier in compilation with improvements to catch more cases
